pyproject.toml
[tool.poetry]
name = "backbone"
version = "0.1.0"
description = ""
authors = ["BalaPradeepR <balapradeepbala@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.143"
botocore = "^1.34.143"
cryptography = "^42.0.8"
fastapi = "^0.111.0"
motor = "^3.5.1"
numpy = "^2.0.0"
pgeocode = "^0.5.0"
pydantic = "^2.8.2"
pytest = "^8.2.2"
requests = "^2.32.3"
sendgrid = "^6.11.0"
starlette = "^0.37.2"
stripe = "^10.2.0"
typing-extensions = "^4.12.2"
uvicorn = "^0.30.1"
firebase-admin = "^6.5.0"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
